+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 ### Added
 
+## [Unreleased]
+### Fixed
+- set collection resource model on all collections items 
+
 ## [101.2.7] - 2023-10-05
 ### Changed
 - Removed Attribute Metadata, it was used in custom project to date format and its not needed anymore