+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed
 - set collection resource model on all collections items 
+- fix argument type error in EavDataSetup
 
 ## [101.2.7] - 2023-10-05
 ### Changed