+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - set collection resource model on all collections items 
 - fix argument type error in EavDataSetup
+### Added
+- serialized backend model
 
 ## [101.2.7] - 2023-10-05
 ### Changed